fix(apply_dataset): 修正数据类型问题并更新配置参数
修正了H5DataGenerator.py中suction_wrench_scores的数据类型转换，确保初始转换为float64类型，并在后续计算中避免数据类型问题。同时，还修正了mapped_values的数据类型转换，避免casting错误。

更新了launch.json中的cycle_list和scene_list参数值，将cycle_list从"[0,10]"改为"9"，将scene_list从"[1,50]"改为"1"。这些更改可能针对特定的调试或测试场景。

+++ apply_dataset/H5DataGenerator.py
@@ -499,7 +499,8 @@ class H5DataGenerator(object):
             torque = np.sqrt(torque_x**2 + torque_y**2)
             score = 1 - min(1, torque / wrench_thre)
             suction_wrench_scores.append(score)
-        suction_wrench_scores = np.array(suction_wrench_scores)
+        # 确保初始转换为float64类型，避免后续计算中的数据类型问题
+        suction_wrench_scores = np.array(suction_wrench_scores, dtype=np.float64)
         
         # 定义参考向量，指向负Z轴方向（垂直向下，符合重力方向）
         reference_vector = np.array([0, 0, -1])
@@ -530,6 +531,9 @@ class H5DataGenerator(object):
         # - 180度（完全垂直向上）→ 权重0.0
         mapped_values = 1 - (angles_degrees / 180)
 
+        # 确保数据类型兼容，避免casting错误
+        suction_wrench_scores = suction_wrench_scores.astype(np.float64)
+        mapped_values = mapped_values.astype(np.float64)
         suction_wrench_scores *= mapped_values # 乘以权重
         return suction_wrench_scores
